导入导出功能重构
