--- lcapDataServer/lcap-server/src/main/java/com/cloudwise/lcap/source/dto/ApplicationsListDto.java
@@ -1,17 +1,0 @@
-package com.cloudwise.lcap.source.dto;
-
-import lombok.*;
-
-import java.io.Serializable;
-import java.util.List;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@ToString
-public class ApplicationsListDto implements Serializable {
-
-    private List<ApplicationViewDto> applications;
-    private int count;
-}
