+++ lcapDataServer/lcap-server/src/main/java/com/cloudwise/lcap/source/service/dto/Manifest.java
@@ -1,9 +1,7 @@
-package com.cloudwise.lcap.source.model;
+package com.cloudwise.lcap.source.service.dto;
 
 import com.cloudwise.lcap.source.dto.ApplicationDto;
-import com.cloudwise.lcap.source.dto.ApplicationsListDto;
 import com.cloudwise.lcap.source.dto.ComponentDto;
-import com.cloudwise.lcap.source.dto.ComponentsListDto;
 import lombok.*;
 
 import java.io.Serializable;
@@ -18,9 +16,6 @@ public class Manifest implements Serializable {
 
     private String applicationExportType;
     private List<String> componentExportType;
-
-    private ApplicationsListDto applications;
-    private ComponentsListDto components;
 
     private String time;
     private String creator;
